Move errors module into separate file
